TCES-23 refactor(employer) change employer info styling to match other pages
