+++ client/src/components/employer-component/employer-info-component/index.jsx
@@ -1,13 +1,15 @@
 // eslint-disable-next-line no-unused-vars
 import * as React from "react";
 import { useState, useRef, useEffect } from "react";
+import { useNavigate } from "react-router-dom";
 import MoreVertIcon from "@mui/icons-material/MoreVert";
 import DownloadIcon from "@mui/icons-material/Download";
+import ArrowBackIcon from "@mui/icons-material/ArrowBack";
 import DeleteIcon from "@mui/icons-material/Delete";
 import { Typography, Menu, MenuItem, IconButton, Box } from "@mui/material";
 
 import PropTypes from "prop-types";
-import { HeaderContainer, TopRowContainer } from "./index.styles";
+import HeaderContainer from "./index.styles";
 import EmployerType from "../../../prop-types/EmployerType";
 import UserType from "../../../prop-types/UserType";
 
@@ -24,6 +26,8 @@ function EmployerInfoComponent({
   const vertButtonRef = useRef(null);
   const [anchorEl, setAnchorEl] = useState(null);
   const open = Boolean(anchorEl);
+
+  const navigate = useNavigate();
 
   let owner = getUserById(employer.owner);
 
@@ -60,69 +64,86 @@ function EmployerInfoComponent({
   if (error) return <ErrorScreenComponent message={error} />;
 
   return (
-    <TopRowContainer
-      style={{ display: "flex", justifyContent: "space-between" }}
-    >
+    <HeaderContainer>
+      <IconButton
+        sx={{
+          marginRight: 2,
+          marginLeft: 2,
+        }}
+        onClick={() => {
+          navigate("/dashboard");
+        }}
+        size="small"
+      >
+        <ArrowBackIcon
+          sx={{
+            color: "gray",
+            cursor: "pointer",
+          }}
+        />
+      </IconButton>
       <div>
-        <HeaderContainer>
-          <Typography
-            style={{
-              fontFamily: "Arial",
-              fontSize: "34px",
-              fontWeight: 500,
-              lineHeight: "42px",
-              letterSpacing: "0.25px",
-              textAlign: "left",
-            }}
-          >
-            {employer.name}
-          </Typography>
-          <IconButton onClick={handleClick}>
-            <MoreVertIcon
-              sx={{
-                color: "gray",
-                cursor: "pointer",
-              }}
-              ref={vertButtonRef}
-            />
-          </IconButton>
-
-          <Menu
-            anchorEl={anchorEl}
-            open={open}
-            onClose={handleClose}
-            MenuListProps={{
-              "aria-labelledby": "split-button",
-            }}
-            PaperProps={{
-              style: {
-                width: vertButtonRef ? 190 : null,
-              },
-            }}
-          >
-            <MenuItem onClick={handleClose}>
-              <IconButton>
-                <DownloadIcon />
-              </IconButton>{" "}
-              Export as JSON
-            </MenuItem>
-            <MenuItem onClick={handleClose}>
-              <IconButton>
-                <DeleteIcon />
-              </IconButton>
-              Delete
-            </MenuItem>
-          </Menu>
-        </HeaderContainer>
         <Typography
           style={{
             fontFamily: "Arial",
+            fontSize: "34px",
+            fontWeight: 500,
+            lineHeight: "42px",
+            letterSpacing: "0.25px",
+            textAlign: "left",
+          }}
+        >
+          {employer.name}
+        </Typography>
+        <Typography
+          style={{
+            fontFamily: "Arial",
+            fontSize: "14px",
+            fontWeight: 400,
+            lineHeight: "24px",
+            letterSpacing: "0.15px",
             textAlign: "left",
           }}
         >
           Employer
         </Typography>
       </div>
+      <IconButton onClick={handleClick}>
+        <MoreVertIcon
+          sx={{
+            color: "gray",
+            cursor: "pointer",
+          }}
+          ref={vertButtonRef}
+        />
+      </IconButton>
+
+      <Menu
+        anchorEl={anchorEl}
+        open={open}
+        onClose={handleClose}
+        MenuListProps={{
+          "aria-labelledby": "split-button",
+        }}
+        PaperProps={{
+          style: {
+            width: vertButtonRef ? 190 : null,
+          },
+        }}
+      >
+        <MenuItem onClick={handleClose}>
+          <IconButton>
+            <DownloadIcon />
+          </IconButton>{" "}
+          Export as JSON
+        </MenuItem>
+        <MenuItem onClick={handleClose}>
+          <IconButton>
+            <DeleteIcon />
+          </IconButton>
+          Delete
+        </MenuItem>
+      </Menu>
 
       <Box
         sx={{
@@ -159,7 +180,7 @@ function EmployerInfoComponent({
         setSnackBarMessage={setSnackBarMessage}
         setError={setError}
       />
-    </TopRowContainer>
+    </HeaderContainer>
   );
 }
