+++ client/src/components/employer-component/employer-info-component/index.styles.jsx
@@ -6,6 +6,4 @@ const HeaderContainer = styled.div`
   margin: 20px;
 `;
 
-const TopRowContainer = styled.div``;
-
-export { HeaderContainer, TopRowContainer };
+export default HeaderContainer;
